Update the rcljava parameter APIs (port osrf/ros2_java#1) (#193)
* Implement all parameter types.

This implementation was missing BoolArray, IntegerArray,
DoubleArray, and StringArray.  Doing this required a change
to the constructor for ParameterVariant, since List is a
Generic and we can't have multiple constructors with the
same signature.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Add log4j initialization to all tests.

It gets rid of a warning when starting the tests that looks like:

log4j:WARN No appenders could be found for logger (org.ros2.rcljava.common.JNIUtils).
log4j:WARN Please initialize the log4j system properly.
log4j:WARN See http://logging.apache.org/log4j/1.2/faq.html#noconfig for more info.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Make the parameter separator a class variable.

It doesn't need to change, so it can just be a class variable.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Store the parameter and the descriptor.

That way we have the information available when other methods
want to retrieve it.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Switch parameters to using map lookups.

The parameters are stored in a HashMap of name -> parameter,
but we were unnecessarily iterating over the HashMap to find
things.  Instead, look up the items directly in the map which
should be much faster.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Plumb undeclaredParameters through the Node constructor.

Whether to allow undeclared parameters is a decision that is
made during Node creation.  Plumb through the necessary option
so that the user can choose to allow undeclared parameters when
they create the node.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Add in another ParameterVariant.

This one just takes a name and sets the type to NOT_SET.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Add in a ParameterCallback interface.

This interface is what users will have to implement in order
to have their callback called when parameters are set.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Update the Node parameter methods to Dashing/Foxy equivalents.

The parameter API was significantly updated during the Dashing
release cycle.  Update the API in rcljava to provide similar
functionality and make the API look a lot more like rclcpp
and rclpy.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Add tests for Node parameters.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Add node parameters test with undeclared parameters.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Fix rebasing error

Signed-off-by: Ivan Santiago Paunovic <ivanpauno@ekumenlabs.com>

* configure logger in an android compatible way

Signed-off-by: Ivan Santiago Paunovic <ivanpauno@ekumenlabs.com>

* configure logger in an android compatible way

Signed-off-by: Ivan Santiago Paunovic <ivanpauno@ekumenlabs.com>

Co-authored-by: Chris Lalancette <clalancette@openrobotics.org>